--- features/stockgetter/src/main/java/com/example/stockgetter/RetrofitClient.kt
@@ -1,18 +1,0 @@
-package com.example.stockgetter
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitClient {
-    private var retrofit: Retrofit? = null
-
-    fun getClient(baseUrl: String): Retrofit {
-        if (retrofit == null) {
-            retrofit = Retrofit.Builder()
-                .baseUrl(baseUrl)
-                .addConverterFactory(GsonConverterFactory.create())
-                .build()
-        }
-        return retrofit!!
-    }
-}
