--- features/stockgetter/src/main/java/com/example/stockgetter/StockApi.kt
@@ -1,10 +1,0 @@
-package com.example.stockgetter
-
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface StockApi {
-    @GET("stock/list")
-    fun getStockList(@Query("apikey") apiKey: String): Call<MutableList<StockModel>>
-}
