--- features/stockgetter/src/main/java/com/example/stockgetter/StockModel.kt
@@ -1,10 +1,0 @@
-package com.example.stockgetter
-
-data class StockModel (
-    val symbol: String?,
-    val name: String?,
-    val price: Double?,
-    val exchange: String?,
-    val exchangeShortName: String?,
-    val type: String?
-)
