--- features/stockgetter/src/main/java/com/example/stockgetter/StockProvider.kt
@@ -1,7 +1,0 @@
-package com.example.stockgetter
-
-object StockProvider {
-    private const val BASE_URL = "https://financialmodelingprep.com/api/v3/"
-    val stockApi: StockApi
-        get() = RetrofitClient.getClient(BASE_URL).create(StockApi::class.java)
-}
