+++ features/stocklist/data/src/main/java/com/example/data/StockApi.kt
@@ -1,0 +1,10 @@
+package com.example.data
+
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface StockApi {
+    @GET("stock/list")
+    fun getStockList(@Query("apikey") apiKey: String): Call<MutableList<StockModel>>
+}
