+++ features/stocklist/data/src/main/java/com/example/data/StockModel.kt
@@ -1,0 +1,10 @@
+package com.example.data
+
+data class StockModel (
+    val symbol: String?,
+    val name: String?,
+    val price: Double?,
+    val exchange: String?,
+    val exchangeShortName: String?,
+    val type: String?
+)
