+++ features/stocklist/data/src/main/java/com/example/data/StockProvider.kt
@@ -1,0 +1,7 @@
+package com.example.data
+
+object StockProvider {
+    private const val BASE_URL = "https://financialmodelingprep.com/api/v3/"
+    val stockApi: StockApi
+        get() = RetrofitClient.getClient(BASE_URL).create(StockApi::class.java)
+}
