+++ features/stocklist/presentation/build.gradle
@@ -1,20 +1,17 @@
 plugins {
-    id 'com.android.application'
+    id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
 }
 
 android {
-    namespace 'com.example.cleanarchitectureshowcase'
+    namespace 'com.example.presentation'
     compileSdk 34
 
     defaultConfig {
-        applicationId "com.example.cleanarchitectureshowcase"
         minSdk 24
-        targetSdk 34
-        versionCode 1
-        versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
     }
 
     buildTypes {
@@ -24,29 +21,26 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_17
-        targetCompatibility JavaVersion.VERSION_17
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
     kotlinOptions {
-        jvmTarget = '17'
+        jvmTarget = '1.8'
     }
 }
 
 dependencies {
-    implementation project(path: ':features:home:presentation')
-    implementation project(path: ':libs:core')
-    implementation project(path: ':features:letsgo')
-    implementation project(path: ':features:stocklist:presentation')
+    implementation project(':features:stocklist:data')
 
-    implementation 'androidx.core:core-ktx:1.12.0'
+    implementation 'androidx.core:core-ktx:1.13.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.11.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-    implementation project(path: ':features:home:domain')
-    implementation project(path: ':features:home:data')
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
     implementation 'androidx.fragment:fragment-ktx:1.6.2'
+    implementation("com.squareup.retrofit2:retrofit:2.9.0")
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.0'
+    implementation 'com.squareup.picasso:picasso:2.71828'
 }
