+++ features/stocklist/presentation/src/main/res/layout/fragment_temp_stock_list.xml
