+++ features/stocklist/presentation/src/main/res/layout/stock_item.xml
