+++ features/stocklist/presentation/src/main/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+
+</resources>
