+++ features/stocklist/src/main/AndroidManifest.xml
