+++ settings.gradle
@@ -20,3 +20,7 @@ include ':features:letsgo'
 include ':features:home:domain'
 include ':features:home:data'
 include ':features:home:presentation'
+include ':features:stocklist'
+include ':features:stocklist:domain'
+include ':features:stocklist:data'
+include ':features:stocklist:presentation'
